Time: 13 ms (66.66%) | Memory: 20.3 MB (47.57%) - LeetSync

+++ 268-missing-number/missing-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int missingNumber(vector<int>& nums) {
-        long long int n=(nums.size()*(nums.size()+1))/2;
+        int n=(nums.size()*(nums.size()+1))/2;
         for(auto i:nums)
            n-=i;
         return n;
